backend/requirements.txt
Django==5.0.0
djangorestframework==3.14.0
django-cors-headers==4.3.1
mysqlclient==2.2.0
djangorestframework-simplejwt==5.3.0
pandas==2.1.3
scikit-learn==1.3.2
python-dotenv==1.0.0